requirements.txt
pandas
numpy
torch
pyannote.audio
scikit-learn
whisper-openai
soundfile
ffmpeg-python
spacy
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.0.0/en_core_web_sm-3.0.0-py3-none-any.whl
